+++ Application/Models/OperationDto.cs
@@ -3,9 +3,9 @@
     public class OperationDto
     {
         public long Id { get; set; }
-        public float FirstValue { get; set; }
-        public float SecondValue { get; set; }
-        public float Result { get; set; }
+        public decimal FirstValue { get; set; }
+        public decimal SecondValue { get; set; }
+        public decimal Result { get; set; }
         public string Operator { get; set; }
     }
 }
