+++ Core/Entities/Operation.cs
@@ -4,9 +4,9 @@
     {
         public long Id { get; set; }
         public string Operator { get; set; }
-        public float FirstValue { get; set; }
-        public float? SecondValue { get; set; }
-        public float Result { get; set; }
+        public decimal FirstValue { get; set; }
+        public decimal? SecondValue { get; set; }
+        public decimal Result { get; set; }
         public DateTimeOffset CreatedAt { get; set; }
     }
 }
