+++ WebApi/Parameters/DoubleOperandsCommandParameters.cs
@@ -2,8 +2,8 @@
 {
     public class DoubleOperandsCommandParameters
     {
-        public float FirstOperand { get; set; }
-        public float SecondOperand { get; set; }
+        public decimal FirstOperand { get; set; }
+        public decimal SecondOperand { get; set; }
         public string Operator { get; set; }
     }
 }
